requirements.txt
beautifulsoup4==4.13.4
numpy==2.2.5
openpyxl==3.1.5
pandas==2.0.3
protobuf==6.30.2
python-dotenv==1.1.0
selenium==4.31.0
sv_ttk==2.6.0
